+++ alf/drivers/experience_replay.py
@@ -17,6 +17,7 @@ import abc
 import tensorflow as tf
 import gin.tf
 from alf.drivers.threads import flatten_once
+from tf_agents.replay_buffers.tf_uniform_replay_buffer import TFUniformReplayBuffer
 
 
 @six.add_metaclass(abc.ABCMeta)
@@ -41,7 +42,7 @@ class ExperienceReplayer(object):
         """
 
     @abc.abstractmethod
-    def replay(self, sample_batch_size, num_steps):
+    def replay(self, sample_batch_size, mini_batch_length):
         """Replay experiences from buffers"""
 
     @abc.abstractmethod
@@ -63,15 +64,14 @@ class OnetimeExperienceReplayer(ExperienceReplayer):
     Example algorithms: IMPALA, PPO2
     """
 
-    def __init__(self):
-        super().__init__()
+    def __init__(self, experience_spec, batch_size):
         self._experience = None
 
     def observe(self, exp, env_ids):
         # flatten the shape (num_envs, env_batch_size)
         self._experience = tf.nest.map_structure(flatten_once, exp)
 
-    def replay(self, sample_batch_size, num_steps):
+    def replay(self, sample_batch_size, mini_batch_length):
         raise NotImplementedError()  # Only supports replaying all!
 
     def replay_all(self):
@@ -79,3 +79,26 @@ class OnetimeExperienceReplayer(ExperienceReplayer):
 
     def clear(self):
         self._experience = None
+
+
+@gin.configurable
+class SyncUniformExperienceReplayer(ExperienceReplayer):
+    """
+    For synchronous off-policy training.
+    """
+
+    def __init__(self, experience_spec, batch_size):
+        self._buffer = TFUniformReplayBuffer(experience_spec, batch_size)
+
+    def observe(self, exp, env_ids=None):
+        self._buffer.add_batch(exp)
+
+    def replay(self, sample_batch_size, mini_batch_length):
+        return self._buffer.get_next(
+            sample_batch_size=sample_batch_size, num_steps=mini_batch_length)
+
+    def replay_all(self):
+        return self._buffer.gather_all()
+
+    def clear(self):
+        self._buffer.clear()
