+++ alf/drivers/off_policy_driver.py
@@ -12,179 +12,82 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import os
-import psutil
-from typing import Callable
-from absl import logging
-from collections import namedtuple
-
-from alf.algorithms.off_policy_algorithm import OffPolicyAlgorithm
-from alf.utils import common
-from alf.algorithms.off_policy_algorithm import Experience, make_experience
-
-import gin.tf
 import tensorflow as tf
 import tensorflow_probability as tfp
 
-from tf_agents.trajectories.trajectory import from_transition
+from alf.drivers import policy_driver
+from alf.utils import common
+
+from alf.algorithms.off_policy_algorithm import OffPolicyAlgorithm, Experience
 from tf_agents.trajectories.policy_step import PolicyStep
 from tf_agents.trajectories.time_step import StepType
-from tf_agents.metrics import tf_metrics
-from tf_agents.utils import eager_utils
+from tf_agents.environments.tf_environment import TFEnvironment
+from alf.algorithms.rl_algorithm import make_training_info
+
 from tf_agents.specs.distribution_spec import DistributionSpec
 from tf_agents.specs.distribution_spec import nested_distributions_from_specs
 
-from alf.drivers.policy_driver import PolicyDriver
-from alf.drivers.threads import TFQueues, ActorThread, EnvThread, LogThread
-from alf.drivers.threads import get_act_dist_param
-from alf.drivers.experience_replay import OnetimeExperienceReplayer
 
-from alf.algorithms.rl_algorithm import make_training_info
-
-
-def transpose2(x, dim1, dim2):
-    perm = list(range(len(x.shape)))
-    perm[dim1] = dim2
-    perm[dim2] = dim1
-    return tf.transpose(x, perm)
-
-
-LearningBatch = namedtuple(
-    "LearningBatch",
-    "time_step policy_step act_dist_param next_time_step env_id")
-
-
-def make_learning_batch(time_step, policy_step, act_dist_param, next_time_step,
-                        env_id):
-    return LearningBatch(
-        time_step=time_step,
-        policy_step=policy_step,
-        act_dist_param=act_dist_param,
-        next_time_step=next_time_step,
-        env_id=env_id)
-
-
-@gin.configurable
-class OffPolicyDriver(PolicyDriver):
+class OffPolicyDriver(policy_driver.PolicyDriver):
     """
-    A driver that enables async training. The definition of 'async' is that prediction
-    and learning are not intervened like below:
-        pred -> learn -> pred -> learn -> ...
-    Instead they are decoupled:
-        pred -> pred -> pred -> ...
-                   |
-            (sync periodically)
-                   |
-        learn -> learn -> learn -> ...
-    And more importantly, a learner or predictor may only operate on a subset of
-    environments at a time.
+    A base class for SyncOffPolicyDriver and AsyncOffPolicyDriver
     """
 
     def __init__(self,
-                 env_f: Callable,
+                 env: TFEnvironment,
                  algorithm: OffPolicyAlgorithm,
-                 num_envs=32,
-                 num_actor_queues=4,
-                 unroll_length=8,
-                 learn_queue_cap=8,
-                 actor_queue_cap=8,
-                 greedy_predict=False,
+                 exp_replayer_class,
                  observers=[],
                  metrics=[],
-                 exp_replayer_class=OnetimeExperienceReplayer,
                  debug_summaries=False,
                  summarize_grads_and_vars=False,
                  train_step_counter=None):
-        """
+        """Create an OffPolicyDriver.
+
         Args:
-            env_f (Callable): a function with 0 args that creates an environment
-            algorithm (OffPolicyAlgorithm):
-            num_envs (int): the number of environments to run asynchronously. Note:
-                    each environment itself could be a tf_agent batched environment.
-                    So the actual total number of environments is `num_envs` * `env_f().batch_size`.
-                    However, `env_f().batch_size` is transparent to this driver. So all
-                    the queues operate on the assumption of `num_envs` environments.
-                    Each environment is exclusively owned by only one `EnvThread`.
-            num_actor_queues (int): number of actor queues. Each queue is exclusivly owned by just one actor
-                            thread.
-            unroll_length (int): number of time steps each environment proceeds before sending the steps
-                            to the learner queue
-            learn_queue_cap (int): the learner queue capacity determines how many environments contribute
-                            to the training data for each training iteration
-            actor_queue_cap (int): the actor queue capacity determines how many environments contribute to
-                            the data for each prediction forward in an `ActorThread`. To prevent deadlock,
-                            it's required that `actor_queue_cap` * `num_actor_queues` <= `num_envs`.
-            greedy_predict (bool): used by the algorithm for prediction
-            observers (list[Callable]): An optional list of observers that are
-                                        updated after every step in the environment. Each observer
-                                        is a callable(time_step.Trajectory).
-            metrics (list[TFStepMetric]): An optiotional list of metrics.
+            env (TFEnvironment): A TFEnvoronmnet
+            algorithm (OffPolicyAlgorithm): The algorithm for training
             exp_replayer_class (ExperienceReplayer): a class that implements how to storie and replay
                                 experiences. See `OnetimeExperienceRelayer` for example. The replayed
                                 experiences are used for parameter updating.
+            observers (list[Callable]): An optional list of observers that are
+                updated after every step in the environment. Each observer is a
+                callable(time_step.Trajectory).
+            metrics (list[TFStepMetric]): An optiotional list of metrics.
             debug_summaries (bool): A bool to gather debug summaries.
             summarize_grads_and_vars (bool): If True, gradient and network
-                                            variable summaries will be written during training.
-            train_step_counter (tf.Variable): An optional counter to increment every time the a new
-                                            iteration is started. If None, it will use
-                                            tf.summary.experimental.get_step(). If this is still None,
-                                            a counter will be created.
+                variable summaries will be written during training.
+            train_step_counter (tf.Variable): An optional counter to increment
+                every time the a new iteration is started. If None, it will use
+                tf.summary.experimental.get_step(). If this is still None, a
+                counter will be created.
         """
         super(OffPolicyDriver, self).__init__(
-            env=env_f(),
+            env=env,
             algorithm=algorithm,
-            observers=[],
-            training=False,
-            greedy_predict=greedy_predict,
+            observers=observers,
+            metrics=metrics,
+            training=False,  # training can only be done by calling self.train()!
+            greedy_predict=False,  # always use OnPolicyDriver for play/eval!
             debug_summaries=debug_summaries,
             summarize_grads_and_vars=summarize_grads_and_vars,
             train_step_counter=train_step_counter)
-        self._coord = tf.train.Coordinator()
-        self._exp_replayer = exp_replayer_class()
+
         self._prepare_specs(algorithm)
         self._trainable_variables = algorithm.trainable_variables
+        self._exp_replayer = exp_replayer_class(self._experience_spec,
+                                                self._env.batch_size)
+        self.add_experience_observer(self._exp_replayer.observe)
 
-        # create actor and env threads
-        self._tfq = TFQueues(
-            num_envs,
-            self._env.batch_size,
-            learn_queue_cap,
-            actor_queue_cap,
-            time_step_spec=self._time_step_spec,
-            policy_step_spec=self._pred_policy_step_spec,
-            act_dist_param_spec=self._action_dist_param_spec,
-            unroll_length=unroll_length,
-            num_actor_queues=num_actor_queues)
-        actor_threads = [
-            ActorThread(
-                name="actor{}".format(i),
-                coord=self._coord,
-                algorithm=self._algorithm,
-                tf_queues=self._tfq,
-                id=i,
-                greedy_predict=greedy_predict,
-                observation_transformer=self._observation_transformer)
-            for i in range(num_actor_queues)
-        ]
-        env_threads = [
-            EnvThread(
-                name="env{}".format(i),
-                coord=self._coord,
-                env_f=env_f,
-                tf_queues=self._tfq,
-                unroll_length=unroll_length,
-                id=i,
-                actor_id=i % num_actor_queues) for i in range(num_envs)
-        ]
-        self._log_thread = LogThread(
-            name="logging",
-            num_envs=num_envs,
-            env_batch_size=self._env.batch_size,
-            observers=observers,
-            metrics=metrics,
-            coord=self._coord,
-            queue=self._tfq.log_queue)
-        self._threads = actor_threads + env_threads + [self._log_thread]
+    @property
+    def exp_replayer(self):
+        return self._exp_replayer
+
+    def start(self):
+        pass
+
+    def stop(self):
+        pass
 
     def _prepare_specs(self, algorithm):
         """Prepare various tensor specs."""
@@ -267,59 +170,6 @@ class OffPolicyDriver(PolicyDriver):
             collect_info=self._processed_experience_spec.info,
             collect_action_distribution=self._action_dist_param_spec)
 
-    @property
-    def exp_replayer(self):
-        return self._exp_replayer
-
-    def get_step_metrics(self):
-        """See PolicyDriver.get_step_metrics()"""
-        return self._log_thread.metrics[:2]
-
-    def get_metrics(self):
-        """See PolicyDriver.get_metrics()"""
-        return self._log_thread.metrics
-
-    def start(self):
-        """Starts all env, actor, and log threads."""
-        for th in self._threads:
-            th.setDaemon(True)
-            th.start()
-        logging.info("All threads started")
-
-    def run(self):
-        """
-        Each call of run() will wait for a learning batch to be filled in by the
-        env threads.
-
-        Output:
-            steps (int): the total number of unrolled steps
-        """
-        t0 = tf.timestamp()
-        batch = make_learning_batch(*self._tfq.learn_queue.dequeue_all())
-        common.summarize_time(name="Learning queue", data=tf.timestamp() - t0)
-        # convert the batch to the experience format
-        exp = make_experience(batch.time_step, batch.policy_step,
-                              batch.act_dist_param)
-        # make the exp batch major for each environment
-        exp = tf.nest.map_structure(lambda e: transpose2(e, 1, 2), exp)
-        self._exp_replayer.observe(exp, batch.env_id)
-
-        num_envs, unroll_length, env_batch_size = batch.time_step.observation.shape[:
-                                                                                    3]
-        return num_envs * unroll_length * env_batch_size
-
-    def stop(self):
-        # finishes the entire program
-        self._coord.request_stop()
-        # Cancel all pending requests (including enqueues and dequeues),
-        # so that no thread hangs before calling coord.should_stop()
-        self._tfq.close_all()
-        self._coord.join(self._threads)
-        logging.info("All threads stopped")
-
-    def _make_time_major(self, nest):
-        return tf.nest.map_structure(lambda x: transpose2(x, 0, 1), nest)
-
     @tf.function
     def train(self,
               experience: Experience,
@@ -376,7 +226,7 @@ class OffPolicyDriver(PolicyDriver):
                 batch = tf.nest.map_structure(
                     lambda x: tf.reshape(x, [mini_batch_size] + list(x.shape)[
                         1:]), batch)
-                batch = self._make_time_major(batch)
+                batch = common.make_time_major(batch)
                 training_info, loss_info, grads_and_vars = self._update(
                     batch, weight=batch.step_type.shape[0] / mini_batch_size)
                 # somehow tf.function autograph does not work correctly for the
@@ -427,7 +277,8 @@ class OffPolicyDriver(PolicyDriver):
                 exp,
                 policy_state,
                 training=True)
-            policy_step, action_dist_param = get_act_dist_param(policy_step)
+            policy_step, action_dist_param = common.get_act_dist_param(
+                policy_step)
 
             training_info = make_training_info(
                 action=exp.action,
