+++ alf/drivers/on_policy_driver.py
@@ -23,15 +23,11 @@ from typing import Callable
 import gin.tf
 import tensorflow as tf
 
-from tf_agents.metrics import tf_metrics
 from tf_agents.environments.tf_environment import TFEnvironment
-from tf_agents.trajectories.trajectory import from_transition
 from tf_agents.specs.distribution_spec import nested_distributions_from_specs
 
 import alf.utils.common as common
 from alf.algorithms.on_policy_algorithm import OnPolicyAlgorithm
-from alf.algorithms.off_policy_algorithm import make_experience
-from alf.algorithms.rl_algorithm import make_action_time_step
 from alf.algorithms.rl_algorithm import make_training_info
 from alf.drivers import policy_driver
 
@@ -104,20 +100,11 @@ class OnPolicyDriver(policy_driver.PolicyDriver):
                 tf.summary.experimental.get_step(). If this is still None, a
                 counter will be created.
         """
-        metric_buf_size = max(10, env.batch_size)
-        standard_metrics = [
-            tf_metrics.NumberOfEpisodes(),
-            tf_metrics.EnvironmentSteps(),
-            tf_metrics.AverageReturnMetric(buffer_size=metric_buf_size),
-            tf_metrics.AverageEpisodeLengthMetric(buffer_size=metric_buf_size),
-        ]
-        self._metrics = standard_metrics + metrics
-        self._exp_observers = []
-
         super(OnPolicyDriver, self).__init__(
             env=env,
             algorithm=algorithm,
-            observers=observers + self._metrics,
+            observers=observers,
+            metrics=metrics,
             training=training,
             greedy_predict=greedy_predict,
             debug_summaries=debug_summaries,
@@ -127,15 +114,9 @@ class OnPolicyDriver(policy_driver.PolicyDriver):
         self._final_step_mode = final_step_mode
 
         if training:
-            self._policy_state_spec = algorithm.train_state_spec
             self._prepare_specs(algorithm)
             self._trainable_variables = algorithm.trainable_variables
             self._train_interval = train_interval
-        else:
-            self._policy_state_spec = algorithm.predict_state_spec
-
-        self._initial_state = common.get_initial_policy_state(
-            self._env.batch_size, self._policy_state_spec)
 
     def _prepare_specs(self, algorithm):
         time_step_spec = self._env.time_step_spec()
@@ -158,108 +139,8 @@ class OnPolicyDriver(policy_driver.PolicyDriver):
             discount=time_step_spec.discount,
             info=info_spec)
 
-    def add_experience_observer(self, observer: Callable):
-        """Add an observer to receive experience.
-
-        Args:
-            observer (Callable): callable which accept Experience as argument.
-        """
-        self._exp_observers.append(observer)
-
-    def get_step_metrics(self):
-        """See PolicyDriver.get_step_metrics()"""
-        return self._metrics[:2]
-
-    def get_metrics(self):
-        """See PolicyDriver.get_metrics()"""
-        return self._metrics
-
-    def get_initial_time_step(self):
-        return common.get_initial_time_step(self._env)
-
-    def get_initial_policy_state(self):
-        return common.get_initial_policy_state(self._env.batch_size,
-                                               self._policy_state_spec)
-
-    def predict(self, max_num_steps, time_step, policy_state):
-        maximum_iterations = math.ceil(max_num_steps / self._env.batch_size)
-        [time_step, policy_state] = tf.while_loop(
-            cond=lambda *_: True,
-            body=self._eval_loop_body,
-            loop_vars=[time_step, policy_state],
-            maximum_iterations=maximum_iterations,
-            back_prop=False,
-            name="predict_loop")
-        return time_step, policy_state
-
-    def _eval_loop_body(self, time_step, policy_state):
-        next_time_step, policy_step, _ = self._step(time_step, policy_state)
-        return [next_time_step, policy_step.state]
-
-    def _step(self, time_step, policy_state):
-        policy_state = common.reset_state_if_necessary(policy_state,
-                                                       self._initial_state,
-                                                       time_step.is_first())
-        policy_step = common.algorithm_step(
-            self._algorithm,
-            self._observation_transformer,
-            time_step,
-            state=policy_state,
-            training=self._training,
-            greedy_predict=self._greedy_predict)
-        action = common.sample_action_distribution(policy_step.action)
-        next_time_step = self._env_step(action)
-        if self._observers:
-            traj = from_transition(time_step,
-                                   policy_step._replace(action=action),
-                                   next_time_step)
-            for observer in self._observers:
-                observer(traj)
-        if self._exp_observers:
-            action_distribution_param = common.get_distribution_params(
-                policy_step.action)
-            exp = make_experience(
-                time_step,
-                policy_step._replace(action=action),
-                action_distribution=action_distribution_param)
-            for observer in self._exp_observers:
-                observer(exp)
-
-        return next_time_step, policy_step, action
-
-    def _env_step(self, action):
-        time_step = self._env.step(action)
-        return make_action_time_step(time_step, action)
-
     @tf.function
-    def run(self, max_num_steps=None, time_step=None, policy_state=None):
-        """
-        Take steps in the environment for max_num_steps.
-
-        If in training mode, algorithm.train_step() and
-        algorithm.train_complete() will be called.
-        If not in training mode, algorithm.predict() will be called.
-
-        The observers will also be called for every environment step.
-
-        Args:
-            max_num_steps (int): stops after so many environment steps. Is the
-                total number of steps from all the individual environment in
-                the bached enviroments including StepType.LAST steps.
-            time_step (ActionTimeStep): optional initial time_step. If None, it
-                will use self.get_initial_time_step(). Elements should be shape
-                [batch_size, ...].
-            policy_state (nested Tensor): optional initial state for the policy.
-
-        Returns:
-            time_step (ActionTimeStep): named tuple with final observation,
-                reward, etc.
-            policy_state (nested Tensor): final step policy state.
-        """
-        if time_step is None:
-            time_step = common.get_initial_time_step(self._env)
-        if policy_state is None:
-            policy_state = self._initial_state
+    def _run(self, max_num_steps, time_step, policy_state):
         if self._training:
             return self.train(max_num_steps, time_step, policy_state)
         else:
