+++ alf/examples/async_driver_ppo_bullet_humanoid.gin
@@ -31,7 +31,15 @@ create_ac_algorithm.actor_fc_layers=(32, 32, 32)
 create_ac_algorithm.value_fc_layers=(32, 32, 32)
 create_ac_algorithm.learning_rate=3e-4
 
+# driver config
+N = 2
+AsyncOffPolicyDriver.num_envs = %N
+AsyncOffPolicyDriver.num_actor_queues = %N
+AsyncOffPolicyDriver.actor_queue_cap = 1
+AsyncOffPolicyDriver.learn_queue_cap = 1
+
 # training config
+off_policy_trainer.train.sync_driver = False
 off_policy_trainer.train.num_iterations = 100000
 off_policy_trainer.train.summarize_grads_and_vars = True
 off_policy_trainer.train.num_updates_per_train_step = 20
